Reorganized entire solution to be dual project for single dll
