+++ RenderSharp/Math/Color/FRGB.cs
@@ -1,6 +1,4 @@
-﻿using RenderSharp.Math;
-
-namespace RenderSharp.RendererCommon
+﻿namespace RenderSharp.Math
 {
     public class FRGB : Vector3<double>
     {
