+++ RenderSharp/Math/Color/HSVA.cs
@@ -1,6 +1,4 @@
-﻿using RenderSharp.Math;
-
-namespace RenderSharp.RendererCommon
+﻿namespace RenderSharp.Math
 {
     public class HSVA : Vector4<double>
     {
