+++ RenderSharp/Math/Vector3.cs
