+++ RenderSharp/Math/VectorOperationException.cs
