+++ RenderSharp/RenderSharp.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Math\Math.csproj" />
-    <ProjectReference Include="..\RendererCommon\RendererCommon.csproj" />
+    <PackageReference Include="Magick.NET-Q16-AnyCPU" Version="13.5.0" />
   </ItemGroup>
 
 </Project>
