--- RenderSharp/RenderSharpExample.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="Renderer\**" />
-    <EmbeddedResource Remove="Renderer\**" />
-    <None Remove="Renderer\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Math\Math.csproj" />
-    <ProjectReference Include="..\Renderer2d\Renderer2d.csproj" />
-    <ProjectReference Include="..\RendererCommon\RendererCommon.csproj" />
-    <ProjectReference Include="..\Scene\Scene.csproj" />
-  </ItemGroup>
-
-</Project>
