--- Renderer2d/Renderer2d.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Math\Math.csproj" />
-    <ProjectReference Include="..\RendererCommon\RendererCommon.csproj" />
-    <ProjectReference Include="..\Scene\Scene.csproj" />
-  </ItemGroup>
-
-</Project>
